requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers==4.40.2
tokenizers<0.20
accelerate>=0.20.0
pillow==11.3.0
requests==2.32.4

networkx>=3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1

websockets>=11.0

numpy>=1.24.0